requirements.txt
asgiref
certifi
charset-normalizer
Django
django-cors-headers
djangorestframework
et-xmlfile
idna
numpy
openpyxl
pandas
psycopg2-binary
python-dateutil
pytz
requests
six
sqlparse
typing_extensions
tzdata
urllib3
